requirements.txt
# StuTally - Dependencies for Week 1 + Week 2
# Python 3.8+

# ==========================================
# Core Framework
# ==========================================
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# ==========================================
# Real-time & WebSockets
# ==========================================
Flask-SocketIO==5.3.5
eventlet==0.33.3
python-socketio==5.10.0

# ==========================================
# Authentication & Security
# ==========================================
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0

# ==========================================
# Computer Vision & AI
# ==========================================
ultralytics>=8.0.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
torch>=2.0.0
torchvision>=0.15.0

# ==========================================
# Object Tracking
# ==========================================
# DeepSort tracking (if using custom tracker)
filterpy==1.4.5
scikit-learn>=1.3.0
scipy>=1.11.0

# ==========================================
# Data Processing
# ==========================================
numpy>=1.24.0,<2.0
pandas>=2.1.0

# ==========================================
# Image Processing
# ==========================================
Pillow>=10.0.0
requests>=2.31.0

# ==========================================
# Data Visualization & Export
# ==========================================
plotly>=5.18.0
matplotlib>=3.8.0
reportlab==4.0.7

# ==========================================
# Database
# ==========================================
# SQLite (built-in with Python)
# For Week 3 PostgreSQL:
# psycopg2-binary==2.9.9
# SQLAlchemy==2.0.23
# alembic==1.13.0

# ==========================================
# Utilities
# ==========================================
python-dateutil>=2.8.2
pytz>=2023.3

# ==========================================
# Development & Testing (optional)
# ==========================================
# pytest>=7.4.3
# pytest-flask>=1.3.0
# pytest-cov>=4.1.0
# black>=23.12.0
# flake8>=6.1.0
# mypy>=1.7.1

# ==========================================
# Production Server (optional)
# ==========================================
# gunicorn==21.2.0
# gevent==23.9.1
# gevent-websocket==0.10.1

# ==========================================
# Monitoring & Logging (optional)
# ==========================================
# psutil==5.9.6

APScheduler>=3.10.4

inference>=0.9.0
